sandbox/pyproject.toml
[project]
name = "sbx"
version = "0.1.0"
description = "Sandbox CLI for the RLM Method pipeline — supports E2B cloud and local Docker backends"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
e2b = ["e2b>=2.6.4"]
docker = []
all = ["e2b>=2.6.4"]

[project.scripts]
sbx = "sbx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
